Refactor TopLevelFilters into own component
