+++ components/UnifiedDocFeed/UnifiedDocFeedMenu.js
@@ -5,18 +5,17 @@ import colors, { pillNavColors, iconColors } from "~/config/themes/colors";
 import FeedOrderingDropdown from "./FeedOrderingDropdown";
 import { feedTypeOpts, topLevelFilters } from "./constants/UnifiedDocFilters";
 import { useRouter } from "next/router";
-import AuthorAvatar from "../AuthorAvatar";
 import { connect } from "react-redux";
 import { getSelectedUrlFilters } from "./utils/getSelectedUrlFilters";
 import MyHubsDropdown from "../Hubs/MyHubsDropdown";
 import handleFilterSelect from "./utils/handleFilterSelect";
 import FeedTab from "./FeedTab";
 import icons from "~/config/themes/icons";
+import TopLevelFilters from "./TopLevelFilters";
 
-const UnifiedDocFeedMenu = ({ currentUser }) => {
+const UnifiedDocFeedMenu = ({}) => {
   const router = useRouter();
 
-  const [isHubSelectOpen, setIsHubSelectOpen] = useState(false);
   const hubsDownRef = useRef(null);
   const [isSmallScreenDropdownOpen, setIsSmallScreenDropdownOpen] =
     useState(false);
@@ -35,10 +34,6 @@ const UnifiedDocFeedMenu = ({ currentUser }) => {
       if (!isTypeFilterClicked) {
         setTagsMenuOpenFor(null);
       }
-
-      console.log("hubsDownRef", hubsDownRef);
-      console.log(e.target);
-      console.log(hubsDownRef.current.contains(e.target));
 
       // if ((hubsDownRef.current.contains(e.target) && isHubSelectOpen) || !hubsDownRef.current.contains(e.target)) {
       //   setIsHubSelectOpen(false);
@@ -118,54 +113,7 @@ const UnifiedDocFeedMenu = ({ currentUser }) => {
     <div className={css(styles.filtersContainer)}>
       <div className={css(styles.buttonGroup)}>
         <div className={css(styles.mainFilters)}>
-          <div className={css(topLevelFilterStyles.container)}>
-            {topLevelFilters.map((f) => (
-              <div
-                className={css(
-                  topLevelFilterStyles.filter,
-                  f.value === selectedFilters.topLevel &&
-                    topLevelFilterStyles.filterSelected
-                )}
-                onClick={() => {
-                  if (f.value === "my-hubs") {
-                    setIsHubSelectOpen(!isHubSelectOpen);
-                  } else {
-                    handleFilterSelect({ router, topLevel: f.value });
-                  }
-                }}
-              >
-                {f.value === "my-hubs" && isHubSelectOpen && <MyHubsDropdown />}
-                <span className={css(topLevelFilterStyles.filterIcon)}>
-                  {f.value === "my-hubs" && (
-                    <AuthorAvatar
-                      author={currentUser?.author_profile}
-                      size={20}
-                    />
-                  )}
-                  {f.icon}
-                </span>
-                <span className={css(topLevelFilterStyles.filterLabel)}>
-                  {f.label}
-                </span>
-                {f.value === "my-hubs" && (
-                  <span
-                    className={css(topLevelFilterStyles.myHubsDown)}
-                    onClick={() => setIsHubSelectOpen(!isHubSelectOpen)}
-                    ref={hubsDownRef}
-                  >
-                    {isHubSelectOpen ? icons.chevronUp : icons.chevronDown}
-                  </span>
-                )}
-                {/* {f.value === "my-hubs" && (
-                  isTagsMenuOpen
-                    ? <span className={css(styles.icon)}>{icons.chevronUp}</span>
-                    : isSelected
-                    ? <span className={css(styles.icon)}>{icons.chevronDown}</span>
-                    : null
-                )} */}
-              </div>
-            ))}
-          </div>
+          <TopLevelFilters selectedFilters={selectedFilters} />
           <div className={css(styles.feedMenu)}>
             <div className={css(styles.filtersAsTabs)}>
               {/* <div className={css(styles.tab, styles.smallScreenFilters)}>
@@ -223,45 +171,6 @@ const UnifiedDocFeedMenu = ({ currentUser }) => {
     </div>
   );
 };
-
-const topLevelFilterStyles = StyleSheet.create({
-  container: {
-    display: "flex",
-    borderBottom: `1px solid ${colors.GREY_LINE(1)}`,
-    width: "100%",
-    marginBottom: 15,
-  },
-  filter: {
-    padding: "0px 4px 12px 0px",
-    display: "flex",
-    position: "relative",
-    marginRight: 25,
-    alignItems: "center",
-    cursor: "pointer",
-    color: colors.BLACK(),
-    ":hover": {
-      color: colors.NEW_BLUE(),
-    },
-  },
-  filterIcon: {
-    marginRight: 8,
-    fontSize: 18,
-  },
-  filterLabel: {},
-  filterSelected: {
-    borderBottom: `2px solid ${colors.NEW_BLUE()}`,
-    color: colors.NEW_BLUE(),
-  },
-  myHubsDown: {
-    marginLeft: 3,
-    padding: "5px 5px",
-    ":hover": {
-      background: iconColors.BACKGROUND,
-      borderRadius: 3,
-      transition: "0.3s",
-    },
-  },
-});
 
 const styles = StyleSheet.create({
   feedMenu: {
@@ -350,8 +259,4 @@ const styles = StyleSheet.create({
   },
 });
 
-const mapStateToProps = (state) => ({
-  currentUser: state.auth.user,
-});
-
-export default connect(mapStateToProps, null)(UnifiedDocFeedMenu);
+export default UnifiedDocFeedMenu;
